+++ refuerzo-web/app/auth.ts
@@ -1,1 +1,2 @@
-export { authOptions, auth } from "./api/auth/[...nextauth]/route"
+export { auth } from "./api/auth/[...nextauth]/route";
+export { authOptions } from "@/lib/auth";
